Initialize render.js
